+++ src/main/kotlin/com/atlassian/performance/tools/awsinfrastructure/virtualusers/AbsentVirtualUsersFormula.kt
@@ -35,7 +35,8 @@ class AbsentVirtualUsers : VirtualUsers {
     override fun applyLoad(
         jira: URI,
         loadProfile: LoadProfile,
-        scenarioClass: Class<out Scenario>?
+        scenarioClass: Class<out Scenario>?,
+        diagnosticsLimit: Int?
     ) {
         logger.debug("Load will not be applied")
     }